scripts/requirements.txt
# Database drivers
psycopg2-binary>=2.9.0  # PostgreSQL driver
